Fixed desync problem with static contexts.
Had to mess with the plumbing, but
overall, I think the fix is way better
than it could have been!

Server and Client now use separate static
contexts for the harvest nodes.  This is
a workaround to make sure that the client/
server don't mix up unreal network IDs,
if someone joins late, and their world
was spawned in a different order.